# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "BSV blockchain indexer API for transaction outputs, blocks, and related data",
        "title": "1Sat Indexer API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/shruggr/1sat-indexer"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "5.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v5/blocks/hash/{hash}": {
            "get": {
                "description": "Get block header information by block hash",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blocks"
                ],
                "summary": "Get block by hash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Block hash",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blk.BlockHeaderResponse"
                        }
                    },
                    "404": {
                        "description": "Block not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v5/blocks/height/{height}": {
            "get": {
                "description": "Get block header information by block height",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blocks"
                ],
                "summary": "Get block by height",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Block height",
                        "name": "height",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blk.BlockHeaderResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid height",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Block not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v5/blocks/list/{from}": {
            "get": {
                "description": "List up to 10,000 block headers starting from a given height",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blocks"
                ],
                "summary": "List blocks",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Starting block height",
                        "name": "from",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/blk.BlockHeaderResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid height",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v5/blocks/tip": {
            "get": {
                "description": "Get the current blockchain tip (highest block)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blocks"
                ],
                "summary": "Get chain tip",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/blk.BlockHeaderResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v5/evt/{tag}/{id}/{value}": {
            "get": {
                "description": "Search for transaction outputs by event tag, id, and value",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get TXOs by event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event tag",
                        "name": "tag",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Event value",
                        "name": "value",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated list of tags to include (use * for all indexed tags)",
                        "name": "tags",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Starting score for pagination",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Reverse order",
                        "name": "rev",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Maximum number of results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include TXO data",
                        "name": "txo",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include script data",
                        "name": "script",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include spend information",
                        "name": "spend",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter for unspent outputs only",
                        "name": "unspent",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/idx.Txo"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "blk.BlockHeaderResponse": {
            "type": "object",
            "properties": {
                "bits": {
                    "type": "integer"
                },
                "creationTimestamp": {
                    "type": "integer"
                },
                "hash": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "merkleRoot": {
                    "type": "string"
                },
                "nonce": {
                    "type": "integer"
                },
                "prevBlockHash": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
        "evt.Event": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "idx.IndexData": {
            "type": "object",
            "properties": {
                "data": {},
                "deps": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/evt.Event"
                    }
                }
            }
        },
        "idx.Txo": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/idx.IndexData"
                    }
                },
                "events": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "height": {
                    "type": "integer"
                },
                "idx": {
                    "type": "integer"
                },
                "outpoint": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "owners": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "satoshis": {
                    "type": "integer"
                },
                "score": {
                    "type": "number"
                },
                "script": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "spend": {
                    "type": "string"
                }
            }
        }
    }
}